Fix OTP verification: type mismatch and UI issues
- Fix OTP code stored as string but column is integer (cast to int)
- Fix OTP verification comparing string to integer (cast to int)
- Fix OTP input UI: white background, visible caret, error display
- Add client-side validation for empty fields
- Add error message display for invalid codes
- Full OTP flow tested end-to-end

+++ app/Http/Controllers/User/Auth/RegisterController.php
@@ -127,7 +127,7 @@ class RegisterController extends Controller
     {
         $data = [
             'user_id'    => $user->id,
-            'code'       => generate_random_code(),
+            'code'       => (int) generate_random_code(),
             'token'      => generate_unique_string('user_authorizations', 'token', 200),
             'created_at' => now(),
         ];
